+++ src/Farmer/Common.fs
@@ -842,10 +842,7 @@ module IPAddressCidr =
     /// Sequence of IP addresses for a CIDR block.
     let addresses (cidr:IPAddressCidr) =
         let first, last = ipRangeNums cidr
-        seq {
-            for i in first..last do
-                yield i |> ofNum
-        }
+        seq { for i in first..last do ofNum i }
     /// Carve a subnet out of an address space.
     let carveAddressSpace (addressSpace:IPAddressCidr) (subnetSizes:int list) =
         let addressSpaceStart, addressSpaceEnd = addressSpace |> ipRangeNums
@@ -866,7 +863,7 @@ module IPAddressCidr =
                         last, cidr
                 if last <= addressSpaceEnd then
                     startAddress <- (last + 1u) |> ofNum
-                    yield cidr
+                    cidr
                 else
                     raise (IndexOutOfRangeException (sprintf "Unable to create subnet %d of /%d" index size))
         }
@@ -1003,3 +1000,6 @@ module Roles =
         let NetworkContributor = RoleID "4d97b98b-1d4f-4787-a291-c67834d212e7"
         let PrivateDnsZoneContributor = RoleID "b12aa53e-6015-4669-85d0-8515ebb3ae7f"
         let TrafficManagerContributor = RoleID "a4b10055-b0c7-44c2-b00f-c7b5b3550cf7"
+
+module Dns =
+    type DnsRecordType = A | AAAA | CName | NS | PTR | TXT | MX